+++ lib/Console/Command/CrawlCommand.php
@@ -3,7 +3,6 @@
 namespace DTL\Extension\Fink\Console\Command;
 
 use Amp\Loop;
-use DTL\Extension\Fink\Console\Display;
 use DTL\Extension\Fink\Console\DisplayBuilder;
 use DTL\Extension\Fink\Console\HeaderParser;
 use DTL\Extension\Fink\Model\DispatcherBuilderFactory;
@@ -103,6 +102,7 @@ class CrawlCommand extends Command
         $this->addOption(self::OPT_HEADER, null, InputOption::VALUE_REQUIRED|InputOption::VALUE_IS_ARRAY, 'Custom header, e.g. "X-Teapot: Me"', []);
         $this->addOption(self::OPT_RATE, null, InputOption::VALUE_REQUIRED, 'Set max request rate (as requests per second)', []);
         $this->addOption(self::OPT_INCLUDE_LINK, null, InputOption::VALUE_REQUIRED|InputOption::VALUE_IS_ARRAY, 'Add an additional URL to the set of URLs under the base URL', []);
+        $this->addOption('display', null, InputOption::VALUE_REQUIRED, 'Display specification, e.g. +memory', '');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -117,7 +117,8 @@ class CrawlCommand extends Command
         });
 
         $section1 = $output->section();
-        $display = $this->displayBuilder->build('');
+
+        $display = $this->displayBuilder->build($this->castToString($input->getOption('display')));
 
         Loop::repeat(self::DISPLAY_POLL_TIME, function () use ($display, $section1, $dispatcher) {
             if (false === $this->shuttingDown) {