+++ lib/FinkExtension.php
@@ -5,7 +5,6 @@ namespace DTL\Extension\Fink;
 use DTL\Extension\Fink\Console\Command\CrawlCommand;
 use DTL\Extension\Fink\Console\DisplayBuilder;
 use DTL\Extension\Fink\Console\DisplayRegistry;
-use DTL\Extension\Fink\Console\Display\ConcatenatingDisplay;
 use DTL\Extension\Fink\Console\Display\MemoryDisplay;
 use DTL\Extension\Fink\Console\Display\RateDisplay;
 use DTL\Extension\Fink\Console\Display\StatusLineDisplay;